server.py moved to a folder
