+++ server/server.py
@@ -1,0 +1,37 @@
+from http.server import HTTPServer, BaseHTTPRequestHandler
+
+class S(BaseHTTPRequestHandler):
+    def _set_headers(self):
+        self.send_response(200)
+        self.send_header("Content-type", "text/html")
+        self.end_headers()
+
+    def _html(self, message):
+        content = "I see:" + message
+        return content.encode("utf8")  # NOTE: must return a bytes object!
+
+    def do_GET(self):
+        self._set_headers()
+        self.wfile.write(self._html("use POST method."))
+
+    def do_HEAD(self):
+        self._set_headers()
+
+    def do_POST(self):
+        # Doesn't do anything with posted data
+        self._set_headers()
+        content_len = int(self.headers.get('Content-Length', 0))
+        post_body = self.rfile.read(content_len)
+        self.wfile.write(self._html(post_body.decode('utf-8')))
+
+
+def run(server_class=HTTPServer, handler_class=S, addr="localhost", port=8000):
+    server_address = (addr, port)
+    httpd = server_class(server_address, handler_class)
+
+    print(f"Starting httpd server on {addr}:{port}")
+    httpd.serve_forever()
+
+if __name__ == "__main__":
+    run()
+
