+++ src/main/kotlin/com/github/bzz/intellij/listeners/MyApplicationActivationListener.kt
