+++ src/main/resources/META-INF/plugin.xml
