+++ src/test/kotlin/com/github/bzz/intellij/MyPluginTest.kt
