+++ src/test/testData/rename/foo.xml
