+++ src/test/testData/rename/foo_after.xml
